requirements.txt
ConfigArgParse==0.14.0
Flask==1.0.2
paramiko==2.4.2
pyserial==3.4
python-chess==0.22.1
#python-chess==0.25.1   (0.25.1 breaks Webserver)
pyOpenSSL==19.0.0
requests==2.21.0
spur==0.3.20
tornado==4.5.1
#tornado==5.1.1   (5.1.1 breaks Webserver - DJ)
configobj==5.0.6
pycrypto==2.6.1
numpy
